requirements.txt
pandas>=1.1.3
numpy>=1.19.2
pybtex>=0.23.0
python-dateutil>=2.8.1

ieee_xplore>=1.0.0
